Experimentation
Current program state runs without player input to move the cue.
I experimented with a variety of methods to no avail.
Cue does not move at all yet I had it moving in a fashion in a clip I sent to Josh. I tried to get it back to that point but still not moving.